refactored internal structure
Moved each BCn implementation into its own module and moved their respective tests with them.
